Fix crash on non-square HATs
The Unicorn pHAT is 8x4, which crashes because it's not a square. The rows index should be offset by the columns, not the rows. I.e. in an 8x4 grid, the pixel index for row=2,col=3 would be row*height + col, or `2 * 4 + 3 = 11`.

Currently the code does row*width + col, or `2 * 8 + 3 = 19` and that will eventually crash!